Update compatibility.md
I can only find anything mentioning this feature [here](http://lua-users.org/lists/lua-l/2012-04/msg00219.html). In Lua 5.1, `__newindex` metamethod is only invoked for `nil` and `NaN` on userdatas; but in Lua 5.2, this has changed that `__newindex` invokes for `nil` and `NaN` on tables as well.

I guess this has compatibility implications and is changing the semantics but I put the status as ' 🤷‍♀️ ' as I'm not sure what zeux's position on this is.

I made a pull request to document this as given that Lua 5.2 has changed a lot with metatables (`__len` on tables, `__pairs` and `__ipairs`, `__lt` and `__le` on different types, and now this), this might as well be intentional.